21.05.07 - 쪽지 작성(7)
1) SQL.xml 인식하는 거 경로가 잘못되어있어 인식을 못했었음. root-context.xml 에서 "mapperLocations"의 value값 변경 (classpath*:/com/project/myver/sql/*SQL.xml
 ->
 classpath:com/project/myver/sql/*.xml)

2) 파일 첨부 안 했을 때 list.isEmty()에 잡히지 않고 for문을 돌아 오류나는 거 수정해야 함!

+++ src/main/java/com/project/myver/service/MemoService.java
@@ -19,8 +19,8 @@ public class MemoService {
 	}
 
 	// 21.05.03 해당 'memo_no' 레코드의 크기 가져오기
-	public int selectRecordSizeFromMemo(int memo_no) {
-		int recordSize = memoDAO.selectRecordSizeFromMemo(memo_no);
+	public double selectRecordSizeFromMemo(int memo_no) {
+		double recordSize = memoDAO.selectRecordSizeFromMemo(memo_no);
 		return recordSize;
 	}
 	
@@ -57,8 +57,8 @@ public class MemoService {
 	}
 
 	// 21.05.03 'memo_file_no'에 해당하는 레코드의 크기 가져오기
-	public int selectRecordSizeFromMemo_file(int memo_file_no) {
-		int recordSize = memoDAO.selectRecordSizeFromMemo_file(memo_file_no);
+	public double selectRecordSizeFromMemo_file(int memo_file_no) {
+		double recordSize = memoDAO.selectRecordSizeFromMemo_file(memo_file_no);
 		return recordSize;
 	}
